Rename Apps.context.js to App.context.js
