+++ src/pages/apps/[id]/environments/[env-id]/remote-config/RemoteConfig.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import Tooltip from "@material-ui/core/Tooltip";
 import { connect } from "~/utils/context";
 import RootContext from "~/pages/Root.context";
-import AppsContext from "~/pages/apps/Apps.context";
+import AppContext from "~/pages/apps/App.context";
 import EnvironmentContext from "~/pages/apps/[id]/environments/[env-id]/Environment.context";
 import Spinner from "~/components/Spinner";
 import InfoBox from "~/components/InfoBox";
@@ -101,7 +101,7 @@ RemoteConfig.propTypes = {
 
 export default connect(RemoteConfig, [
   { Context: RootContext, props: ["api"] },
-  { Context: AppsContext, props: ["app"] },
+  { Context: AppContext, props: ["app"] },
   { Context: EnvironmentContext, props: ["environment"] },
   { Context: ParameterModal, props: ["toggleModal"], wrap: true },
 ]);
