+++ src/pages/apps/[id]/environments/[env-id]/remote-config/_components/ParameterModal.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import cn from "classnames";
 import { withRouter } from "react-router-dom";
 import RootContext from "~/pages/Root.context";
-import AppsContext from "~/pages/apps/Apps.context";
+import AppContext from "~/pages/apps/App.context";
 import EnvironmentContext from "~/pages/apps/[id]/environments/[env-id]/Environment.context";
 import Modal from "~/components/Modal";
 import Form from "~/components/Form";
@@ -191,7 +191,7 @@ export default Object.assign(
   connect(withRouter(ParameterModal), [
     { Context: ModalContext, props: ["toggleModal", "isOpen"] },
     { Context: RootContext, props: ["api"] },
-    { Context: AppsContext, props: ["app"] },
+    { Context: AppContext, props: ["app"] },
     { Context: EnvironmentContext, props: ["environment"] },
   ]),
   ModalContext
