+++ src/pages/apps/[id]/environments/[env-id]/snippets/_components/SnippetModal.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import { withRouter } from "react-router-dom";
 import RootContext from "~/pages/Root.context";
-import AppsContext from "~/pages/apps/Apps.context";
+import AppContext from "~/pages/apps/App.context";
 import EnvironmentContext from "~/pages/apps/[id]/environments/[env-id]/Environment.context";
 import Modal from "~/components/Modal";
 import Form from "~/components/Form";
@@ -164,7 +164,7 @@ export default Object.assign(
   connect(withRouter(SnippetModal), [
     { Context: ModalContext, props: ["toggleModal", "isOpen"] },
     { Context: RootContext, props: ["api"] },
-    { Context: AppsContext, props: ["app"] },
+    { Context: AppContext, props: ["app"] },
     { Context: EnvironmentContext, props: ["environment"] },
   ]),
   ModalContext
